refactor: revert "refactor(email-submission): encapsulate parsedResponses (#2206)" (#2248)
This reverts commit 6512a84c6456665e6f12a8fad5a7e8ccd9860d28.
