+++ src/app/modules/form/admin-form/__tests__/admin-form.controller.spec.ts
@@ -77,7 +77,6 @@ import {
 } from 'tests/unit/backend/helpers/generate-form-data'
 import expressHandler from 'tests/unit/backend/helpers/jest-express'
 
-import ParsedResponsesObject from '../../../submission/email-submission/ParsedResponsesObject.class'
 import * as UserService from '../../../user/user.service'
 import {
   ForbiddenFormError,
@@ -128,8 +127,6 @@ jest.mock('src/app/modules/submission/submission.utils')
 const MockSubmissionUtils = mocked(SubmissionUtils)
 jest.mock('../admin-form.service')
 const MockAdminFormService = mocked(AdminFormService)
-jest.mock('../../../submission/email-submission/ParsedResponsesObject.class')
-const MockParsedResponsesObject = mocked(ParsedResponsesObject)
 jest.mock('../../form.service')
 const MockFormService = mocked(FormService)
 jest.mock('../../../user/user.service')
@@ -5133,12 +5130,8 @@ describe('admin-form.controller', () => {
       MockEmailSubmissionService.validateAttachments.mockReturnValue(
         okAsync(true),
       )
-      MockParsedResponsesObject.mockClear()
-      MockParsedResponsesObject.parseResponses.mockReturnValue(
-        ok(new ParsedResponsesObject(MOCK_PARSED_RESPONSES)),
-      )
-      MockParsedResponsesObject.mock.instances[0].getAllResponses.mockReturnValue(
-        MOCK_PARSED_RESPONSES,
+      MockSubmissionService.getProcessedResponses.mockReturnValue(
+        ok(MOCK_PARSED_RESPONSES),
       )
       MockEmailSubmissionService.createEmailSubmissionWithoutSave.mockReturnValue(
         MOCK_SUBMISSION,
@@ -5193,7 +5186,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).toHaveBeenCalledWith(
+      expect(MockSubmissionService.getProcessedResponses).toHaveBeenCalledWith(
         MOCK_FORM,
         MOCK_RESPONSES,
       )
@@ -5260,7 +5253,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).not.toHaveBeenCalled()
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5309,7 +5302,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).not.toHaveBeenCalled()
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5362,7 +5355,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).not.toHaveBeenCalled()
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5415,7 +5408,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).not.toHaveBeenCalled()
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5468,7 +5461,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).not.toHaveBeenCalled()
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5521,7 +5514,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).not.toHaveBeenCalled()
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5574,7 +5567,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).not.toHaveBeenCalled()
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5627,7 +5620,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5680,7 +5673,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).not.toHaveBeenCalled()
+      expect(MockSubmissionService.getProcessedResponses).not.toHaveBeenCalled()
       expect(MockAdminFormService.extractMyInfoFieldIds).not.toHaveBeenCalled()
       expect(
         MockEmailSubmissionService.createEmailSubmissionWithoutSave,
@@ -5699,7 +5692,7 @@ describe('admin-form.controller', () => {
     })
 
     it('should return 400 when responses cannot be processed', async () => {
-      MockParsedResponsesObject.parseResponses.mockReturnValueOnce(
+      MockSubmissionService.getProcessedResponses.mockReturnValueOnce(
         err(new ProcessingError()),
       )
       const mockReq = expressHandler.mockRequest({
@@ -5733,7 +5726,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).toHaveBeenCalledWith(
+      expect(MockSubmissionService.getProcessedResponses).toHaveBeenCalledWith(
         MOCK_FORM,
         MOCK_RESPONSES,
       )
@@ -5755,7 +5748,7 @@ describe('admin-form.controller', () => {
     })
 
     it('should return 409 when the submitted field IDs do not match the form field IDs', async () => {
-      MockParsedResponsesObject.parseResponses.mockReturnValueOnce(
+      MockSubmissionService.getProcessedResponses.mockReturnValueOnce(
         err(new ConflictError('')),
       )
       const mockReq = expressHandler.mockRequest({
@@ -5789,7 +5782,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).toHaveBeenCalledWith(
+      expect(MockSubmissionService.getProcessedResponses).toHaveBeenCalledWith(
         MOCK_FORM,
         MOCK_RESPONSES,
       )
@@ -5811,7 +5804,7 @@ describe('admin-form.controller', () => {
     })
 
     it('should return 400 when any answer is invalid', async () => {
-      MockParsedResponsesObject.parseResponses.mockReturnValueOnce(
+      MockSubmissionService.getProcessedResponses.mockReturnValueOnce(
         err(new ValidateFieldError()),
       )
       const mockReq = expressHandler.mockRequest({
@@ -5845,7 +5838,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).toHaveBeenCalledWith(
+      expect(MockSubmissionService.getProcessedResponses).toHaveBeenCalledWith(
         MOCK_FORM,
         MOCK_RESPONSES,
       )
@@ -5901,7 +5894,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).toHaveBeenCalledWith(
+      expect(MockSubmissionService.getProcessedResponses).toHaveBeenCalledWith(
         MOCK_FORM,
         MOCK_RESPONSES,
       )
@@ -5966,7 +5959,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).toHaveBeenCalledWith(
+      expect(MockSubmissionService.getProcessedResponses).toHaveBeenCalledWith(
         MOCK_FORM,
         MOCK_RESPONSES,
       )
@@ -6031,7 +6024,7 @@ describe('admin-form.controller', () => {
       expect(
         MockEmailSubmissionService.validateAttachments,
       ).toHaveBeenCalledWith(MOCK_RESPONSES)
-      expect(MockParsedResponsesObject.parseResponses).toHaveBeenCalledWith(
+      expect(MockSubmissionService.getProcessedResponses).toHaveBeenCalledWith(
         MOCK_FORM,
         MOCK_RESPONSES,
       )
